# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Delivery API",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o token JWT no formato: Bearer {token}"
    }
  },
...
  "paths": {
...
    "/payment-methods": {
      "get": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Listar meios de pagamento do usuário",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de cartões"
          }
        }
      },
      "post": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Cadastrar novo cartão",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "cartao",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nomeTitular": {
                  "type": "string"
                },
                "numeroCartao": {
                  "type": "string"
                },
                "validade": {
                  "type": "string"
                },
                "cvv": {
                  "type": "string"
                },
                "tipo": {
                  "type": "string",
                  "enum": [
                    "crédito",
                    "débito"
                  ]
                }
              },
              "required": [
                "nomeTitular",
                "numeroCartao",
                "validade",
                "cvv",
                "tipo"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Cartão salvo com sucesso"
          }
        }
      }
    },
    "/payment-methods/{id}": {
      "patch": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Atualizar cartão",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "dados",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cartão atualizado"
          }
        }
      },
      "delete": {
        "tags": [
          "Pagamentos"
        ],
        "summary": "Excluir cartão",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Cartão removido com sucesso"
          }
        }
      }
    },
...
    "/orders": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Criar novo pedido",
        "description": "Cria um novo pedido para a empresa especificada. Usuários bloqueados pela empresa não poderão realizar pedidos.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "pedido",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "companyId": {
                  "type": "integer",
                  "example": 123
                },
                "addressId": {
                  "type": "string",
                  "example": "end-456"
                },
                "items": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "productId": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      },
                      "quantity": {
                        "type": "integer"
                      },
                      "observation": {
                        "type": "string"
                      },
                      "addOns": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "price": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    },
                    "required": [
                      "productId",
                      "name",
                      "price",
                      "quantity"
                    ]
                  }
                }
              },
              "required": [
                "companyId",
                "addressId",
                "items"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Pedido criado com sucesso"
          },
          "400": {
            "description": "Campos obrigatórios faltando"
          },
          "403": {
            "description": "Usuário bloqueado pela empresa"
          },
          "404": {
            "description": "Empresa não encontrada"
          }
        }
      },
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Listar pedidos do usuário logado",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos"
          }
        }
      }
    },
    "/orders/active": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Buscar pedido ativo atual do usuário",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido em andamento retornado com sucesso"
          },
          "204": {
            "description": "Nenhum pedido em andamento encontrado"
          }
        }
      }
    },
    "/orders/{id}/details": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Detalhes completos do pedido para o usuário logado",
        "description": "Retorna todas as informações de um pedido específico, incluindo histórico de status, itens, empresa e entregador (se houver)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do pedido"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes completos do pedido",
            "schema": {
              "type": "object",
              "properties": {
                "numeroPedido": {
                  "type": "string"
                },
                "data": {
                  "type": "string",
                  "format": "date-time"
                },
                "statusAtual": {
                  "type": "string"
                },
                "statusHistorico": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "total": {
                  "type": "number"
                },
                "itens": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "nome": {
                        "type": "string"
                      },
                      "quantidade": {
                        "type": "integer"
                      },
                      "preco": {
                        "type": "number"
                      },
                      "observacao": {
                        "type": "string"
                      },
                      "adicionais": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "price": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "empresa": {
                  "type": "object",
                  "properties": {
                    "nome": {
                      "type": "string"
                    },
                    "imagem": {
                      "type": "string"
                    }
                  }
                },
                "entregador": {
                  "type": "object",
                  "description": "Informações do entregador atribuído (se houver)",
                  "properties": {
                    "nome": {
                      "type": "string"
                    },
                    "placa": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      }
    },
    "/orders/{id}/cancel": {
      "patch": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Cancelar pedido (cliente)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido cancelado com sucesso"
          },
          "400": {
            "description": "Não é possível cancelar"
          },
          "404": {
            "description": "Pedido não encontrado"
          }
        }
      }
    },
    "/orders/{id}/repeat": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Repetir um pedido anterior (jogar no carrinho)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do pedido a ser repetido"
          }
        ],
        "responses": {
          "201": {
            "description": "Carrinho atualizado com os itens do pedido anterior"
          },
          "404": {
            "description": "Pedido não encontrado ou não pertence ao usuário"
          }
        }
      }
    },
    "/orders/repeat-last": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Repetir automaticamente o último pedido entregue",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Carrinho preenchido com itens do último pedido"
          },
          "404": {
            "description": "Nenhum pedido entregue encontrado"
          }
        }
      }
    },
...
  }
}